+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2022.1.2 (WIP)
 * rework Babel icon, and provide an alternative icon.
+* support more Prettier config files.
 
 ## 2022.1.1 (2022/03/05)
 * **INFO**: Extra Icons is now a paid plugin: 5$ per year. Price decreases over time, and it's free for students, teachers and OSS developers. If you don't want to (or can't) support my work, feel free to use [previous releases](https://github.com/jonathanlermitage/intellij-extra-icons-plugin/releases/tag/v1.69).